Change LoadOperation and StoreOperation to use MemRef dialect
Currently, LoadOperation and StoreOperation expect the instructions to
just be "load" and "store", respectively. This is not possible in
current MLIR - indeed, I don't know if it was ever possible. Regardless,
since the operations involve MemRef types anyway, this isn't a
significant change, and just makes it compliant with modern MLIR.

Specifically, this change adds the MemRef dialect, whose currently
supported operations are `load` and `store`. (more to come!)